requirements.txt
appdirs==1.4.3
Beaker==1.8.1
cachetools==2.0.1
chardet==3.0.4
cupshelpers==1.0
dill==0.2.7.1
docopt==0.6.2
future==0.16.0
google-auth==1.1.1
google-cloud-core==0.27.1
google-cloud-firestore==0.27.0
google-gax==0.15.15
googleapis-common-protos==1.5.3
grpcio==1.4.0
idna==2.6
keyutils==0.5
louis==3.3.0
Mako==1.0.7
MarkupSafe==1.0
mugshot==0.3.2
nose==1.3.7
npyscreen==4.10.5
packaging==16.8
pacman-mirrors==4.2.3
pexpect==4.2.1
ply==3.8
protobuf==3.4.0
psutil==5.3.1
ptyprocess==0.5.2
py==1.4.34
pyasn1==0.3.7
pyasn1-modules==0.1.5
pycairo==1.13.3
pycups==1.9.73
pycurl==7.43.0
pygobject==3.24.1
pyparsing==2.2.0
pytest==3.2.3
python-distutils-extra==2.39
python-xapp==1.0.0
PyYAML==3.12
requests==2.18.4
rsa==3.4.2
setproctitle==1.1.10
six==1.11.0
team==1.0
udiskie==1.7.1
urllib3==1.22
virtualenv==15.1.0
